requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.60.1
django-bootstrap4==24.1
django-elasticsearch-dsl==8.0
django-environ==0.11.2
djangorestframework==3.14.0
elastic-transport==8.12.0
elasticsearch==8.12.0
elasticsearch-dsl==8.12.0
gunicorn==21.2.0
idna==3.6
isodate==0.6.1
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.0
